+++ src/dstparser/cli/slurm.py
@@ -1,19 +1,7 @@
+import os
 import subprocess
 import shutil
 from pathlib import Path
-
-
-def slurm_directives():
-    return {
-        "job-name": "procsh",
-        "array": "0",
-        "ntasks": 1,
-        # "exclude": "hpa-wn[11,13]",
-        "mem": "20gb",
-        "cpus-per-task": 1,
-        "partition": "short_serial",
-        "time": "01:00:00",
-    }
 
 
 def copy_files(src_paths, dest_dir):
@@ -35,6 +23,42 @@ def copy_files(src_paths, dest_dir):
             shutil.copyfile(file, str(dest_dir / file.name))
 
 
+def slurm_parameters():
+
+    env_vars = {
+        "task_id": "SLURM_ARRAY_TASK_ID",
+        "task_min": "SLURM_ARRAY_TASK_MIN",
+        "task_max": "SLURM_ARRAY_TASK_MAX",
+        "job_id": "SLURM_JOB_ID",
+    }
+
+    slurm_params = dict()
+
+    for key, envar in env_vars.items():
+        val = os.environ.get(envar)
+        try:
+            slurm_params[key] = int(val)
+        except TypeError:
+            slurm_params[key] = val
+
+    return slurm_params
+
+
+def task_info(zero_indexing=False):
+    slurm_params = slurm_parameters()
+
+    if slurm_params["task_id"] is None:
+        return None, None
+
+    task_id = slurm_params["task_id"]
+    ntasks = slurm_params["task_max"] - slurm_params["task_min"] + 1
+
+    if zero_indexing:
+        task_id -= slurm_params["task_min"]
+
+    return task_id, ntasks
+
+
 def run_slurm_job(
     slurm_directives,
     log_dir,
@@ -44,6 +68,21 @@ def run_slurm_job(
     suffix="",
     batch_command="sbatch",
 ):
+    """
+    slurm_directives is dictionary, e.g.:
+
+    slurm_directives = {
+        "job-name": "procsh",
+        "array": "0",
+        "ntasks": 1,
+        "exclude": "hpa-wn[11,13]",
+        "mem": "20gb",
+        "cpus-per-task": 1,
+        "partition": "hdr1-al9_short",
+        "time": "01:00:00",
+    }
+    """
+
     # Create log dirs
     log_dir = Path(log_dir)
     out_dir = log_dir / "output"
@@ -59,6 +98,7 @@ def run_slurm_job(
     slurm_script += f"{sbatch}output={str(out_dir)}/%x-%A_%4a.out"
     slurm_script += f"{sbatch}error={str(err_dir)}/%x-%A_%4a.err"
 
+    # cd to the directory of launching script
     working_dir = Path(script).parent
     command = f"{app.strip()} {str(script).strip()} {options.strip()}"
     slurm_script += f"\n\ncd {str(working_dir)}\n{command.strip()}"
@@ -74,53 +114,8 @@ def run_slurm_job(
         [batch_command, f"{slurm_file}"], stdout=subprocess.PIPE, stderr=subprocess.PIPE
     )
 
+    # Print whether job was submitted
     print(proc.stdout.decode())
     err = proc.stderr.decode()
     if err.strip():
         print(f"stderr:\n{err}")
-
-
-def run_main_job():
-
-    output_dir = (
-        "/ceph/work/SATORI/projects/TA-ASIoP/dnn_training_data/2024/08/01_mc_wdate/"
-    )
-
-    data_dirs = [
-        # "/ceph/work/SATORI/projects/TA-ASIoP/tasdmc_dstbank/qgsii04proton/080417_160603/Em1_bsdinfo",
-        # "/ceph/work/SATORI/projects/TA-ASIoP/tasdmc_dstbank/qgsii04iron/080417_160603/Em1_bsdinfo",
-        "/ceph/work/SATORI/projects/TA-ASIoP/tasdmc_dstbank/qgsii04helium/080417_160603/Em1_bsdinfo",
-    ]
-
-    # data_dirs = [
-    #     "/ceph/work/SATORI/projects/TA-ASIoP/tasdobs_dstbank/rufldf",
-    # ]
-
-    data_dirs = " ".join(data_dirs)
-    # patterns = "tasdcalibev*rufldf.dst.gz"
-    patterns = "DAT*dst.gz"
-    num_temp_files = 1000
-    num_final_files = 50
-
-    script = Path(__file__).parent / "main_job.py"
-    log_dir = Path(output_dir) / "slurm_run"
-    options = f"--log_dir {str(log_dir)} "
-    options += f"--data_dirs {data_dirs} "
-    options += f'--glob_patterns "{patterns}" '
-    options += f"--output_dir {output_dir} "
-    options += f"--temp_h5_files {num_temp_files} "
-    options += f"--final_h5_files {num_final_files} "
-
-    # print(options)
-    run_slurm_job(
-        slurm_directives(),
-        log_dir,
-        script,
-        options,
-        suffix="_main_job",
-        batch_command="bash",
-    )
-
-
-if __name__ == "__main__":
-    run_main_job()
