+++ src/dstparser/dst_adapter.py
@@ -2,6 +2,9 @@ import numpy as np
 from dstparser.dst_reader import read_dst_file
 from dstparser.dst_parsers import parse_dst_string
 import dstparser.tasd_clf as tasd_clf
+from dstparser.event_ids import add_event_ids
+import re
+from pathlib import Path
 
 
 def corsika_id2mass(corsika_pid):
@@ -360,6 +363,7 @@ def parse_dst_file(
     avg_traces=True,
     add_shower_params=True,
     add_standard_recon=True,
+    # add_event_id=
 ):
     #  ntile - number of SD per one side
     dst_string = read_dst_file(dst_file)
@@ -381,4 +385,7 @@ def parse_dst_file(
         data = standard_recon(data, dst_lists)
 
     data = detector_readings(data, dst_lists, ntile, avg_traces)
+
+    if add_event_ids:
+        data = add_event_ids(data, dst_file)
     return data
