+++ tests/test_parser.py
@@ -34,8 +34,11 @@ def test_parser(dst_file, print_read_data=False):
     # print(f'energy = {data["energy"]}')
     # print(f'xmax = {data["xmax"]}')
     # print(f'time_traces_low = {data["time_traces_low"]}')
+    # print(f'event id = {data["event_id"]}')
+    # print(f'corsika shower id = {data["corsika_shower_id"]}')
+    # print(f'energy bin id = {data["energy_bin_id"]}')
 
 
 if __name__ == "__main__":
-    dst_file = "/ceph/work/SATORI/projects/TA-ASIoP/tasdmc_dstbank/qgsii04proton/080417_160603/Em1_bsdinfo/DAT000007_gea.rufldf.dst.gz"
-    test_parser(dst_file, print_read_data=False)
+    dst_file = "/ceph/work/SATORI/projects/TA-ASIoP/tasdmc_dstbank/qgsii04proton/080417_160603/Em1_bsdinfo/DAT081325_gea.rufldf.dst.gz"
+    test_parser(dst_file, print_read_data=True)
